--- web/src/pages/createPlaylist.js
@@ -1,81 +1,0 @@
-import MusicPlaylistClient from '../api/musicPlaylistClient';
-import Header from '../components/header';
-import BindingClass from '../util/bindingClass';
-import DataStore from '../util/DataStore';
-
-/**
- * Logic needed for the create playlist page of the website.
- */
-class CreatePlaylist extends BindingClass {
-    constructor() {
-        super();
-        this.bindClassMethods(['mount', 'submit', 'redirectToViewPlaylist'], this);
-        this.dataStore = new DataStore();
-        this.dataStore.addChangeListener(this.redirectToViewPlaylist);
-        this.header = new Header(this.dataStore);
-    }
-
-    /**
-     * Add the header to the page and load the MusicPlaylistClient.
-     */
-    mount() {
-        document.getElementById('create').addEventListener('click', this.submit);
-
-        this.header.addHeaderToPage();
-
-        this.client = new MusicPlaylistClient();
-    }
-
-    /**
-     * Method to run when the create playlist submit button is pressed. Call the MusicPlaylistService to create the
-     * playlist.
-     */
-    async submit(evt) {
-        evt.preventDefault();
-
-        const errorMessageDisplay = document.getElementById('error-message');
-        errorMessageDisplay.innerText = ``;
-        errorMessageDisplay.classList.add('hidden');
-
-        const createButton = document.getElementById('create');
-        const origButtonText = createButton.innerText;
-        createButton.innerText = 'Loading...';
-
-        const playlistName = document.getElementById('playlist-name').value;
-        const tagsText = document.getElementById('tags').value;
-
-        let tags;
-        if (tagsText.length < 1) {
-            tags = null;
-        } else {
-            tags = tagsText.split(/\s*,\s*/);
-        }
-
-        const playlist = await this.client.createPlaylist(playlistName, tags, (error) => {
-            createButton.innerText = origButtonText;
-            errorMessageDisplay.innerText = `Error: ${error.message}`;
-            errorMessageDisplay.classList.remove('hidden');
-        });
-        this.dataStore.set('playlist', playlist);
-    }
-
-    /**
-     * When the playlist is updated in the datastore, redirect to the view playlist page.
-     */
-    redirectToViewPlaylist() {
-        const playlist = this.dataStore.get('playlist');
-        if (playlist != null) {
-            window.location.href = `/playlist.html?id=${playlist.id}`;
-        }
-    }
-}
-
-/**
- * Main method to run when the page contents have loaded.
- */
-const main = async () => {
-    const createPlaylist = new CreatePlaylist();
-    createPlaylist.mount();
-};
-
-window.addEventListener('DOMContentLoaded', main);
