+++ web/src/pages/searchPrescriptions.js
@@ -3,17 +3,6 @@ import Header from '../components/header';
 import BindingClass from "../util/bindingClass";
 import DataStore from "../util/DataStore";
 
-/*
-The code below this comment is equivalent to...
-const EMPTY_DATASTORE_STATE = {
-    'search-criteria': '',
-    'search-results': [],
-};
-
-...but uses the "KEY" constants instead of "magic strings".
-The "KEY" constants will be reused a few times below.
-*/
-
 const SEARCH_CRITERIA_KEY = 'search-criteria';
 const SEARCH_RESULTS_KEY = 'search-results';
 const EMPTY_DATASTORE_STATE = {
@@ -21,21 +10,17 @@ const EMPTY_DATASTORE_STATE = {
     [SEARCH_RESULTS_KEY]: [],
 };
 
-
 /**
  * Logic needed for the view playlist page of the website.
  */
 class SearchPrescriptions extends BindingClass {
     constructor() {
         super();
-
         this.bindClassMethods(['mount', 'search', 'displaySearchResults', 'getHTMLForSearchResults', 'redirectToViewPrescription'], this);
 
-        // Create a enw datastore with an initial "empty" state.
         this.dataStore = new DataStore(EMPTY_DATASTORE_STATE);
         this.header = new Header(this.dataStore);
         this.dataStore.addChangeListener(this.displaySearchResults);
-//        this.dataStore.addChangeListener(this.redirectToViewPrescription);
 
         console.log("searchPrescriptions constructor");
     }
@@ -47,11 +32,9 @@ class SearchPrescriptions extends BindingClass {
         // Wire up the form's 'submit' event and the button's 'click' event to the search method.
         document.getElementById('search-prescriptions-form').addEventListener('submit', this.search);
         document.getElementById('search-btn').addEventListener('click', this.search);
-//        document.getElementById('search-prescriptions-form').addEventListener('submit', this.search);
         document.getElementById('search-single-btn').addEventListener('click', this.redirectToViewPrescription);
 
         this.header.addHeaderToPage();
-
         this.client = new PharmacyClient();
     }
 
@@ -68,7 +51,6 @@ class SearchPrescriptions extends BindingClass {
         let status = document.getElementById('status').value;
         console.log(email);
         console.log(status);
-
 
         if (email === '') {
             email = 'none';
@@ -92,7 +74,8 @@ class SearchPrescriptions extends BindingClass {
                 [SEARCH_CRITERIA_KEY]: searchCriteria,
                 [SEARCH_RESULTS_KEY]: results,
             });
-        } else {
+        }
+        else {
             this.dataStore.setState(EMPTY_DATASTORE_STATE);
         }
     }
