+++ web/src/pages/viewPatient.js
@@ -13,7 +13,7 @@ class ViewPatient extends BindingClass {
         this.dataStore = new DataStore();
         this.dataStore.addChangeListener(this.addPatientToPage);
         this.header = new Header(this.dataStore);
-        console.log("viewpatient constructor");
+        console.log("viewPatient constructor");
     }
 
     /**
@@ -22,8 +22,8 @@ class ViewPatient extends BindingClass {
     async clientLoaded() {
         const urlParams = new URLSearchParams(window.location.search);
         const email = urlParams.get('email');
-//        document.getElementById('patient').innerText = "Loading Patient ...";
         const patient = await this.client.viewPatient(email);
+
         this.dataStore.set('patient', patient);
     }
 
@@ -31,8 +31,6 @@ class ViewPatient extends BindingClass {
      * Add the header to the page and load the PharmacyClient.
      */
     mount() {
-//        document.getElementById('search-single-btn').addEventListener('click', this.addPatientToPage);
-
         this.header.addHeaderToPage();
 
         this.client = new PharmacyClient();
@@ -44,6 +42,7 @@ class ViewPatient extends BindingClass {
      */
     addPatientToPage() {
         const patient = this.dataStore.get('patient');
+
         if (patient == null) {
             return;
         }
@@ -55,60 +54,6 @@ class ViewPatient extends BindingClass {
         document.getElementById('phone').value = patient.phone;
         document.getElementById('address').value = patient.address;
     }
-
-    /**
-     * When the songs are updated in the datastore, update the list of songs on the page.
-     */
-//    addSongsToPage() {
-//        const songs = this.dataStore.get('songs')
-//
-//        if (songs == null) {
-//            return;
-//        }
-//
-//        let songHtml = '';
-//        let song;
-//        for (song of songs) {
-//            songHtml += `
-//                <li class="song">
-//                    <span class="title">${song.title}</span>
-//                    <span class="album">${song.album}</span>
-//                </li>
-//            `;
-//        }
-//        document.getElementById('songs').innerHTML = songHtml;
-//    }
-
-    /**
-     * Method to run when the add song patient submit button is pressed. Call the MusicPlaylistService to add a song to the
-     * patient.
-     */
-//    async addSong() {
-//
-//        const errorMessageDisplay = document.getElementById('error-message');
-//        errorMessageDisplay.innerText = ``;
-//        errorMessageDisplay.classList.add('hidden');
-//
-//        const patient = this.dataStore.get('patient');
-//        if (patient == null) {
-//            return;
-//        }
-//
-//        document.getElementById('add-song').innerText = 'Adding...';
-//        const asin = document.getElementById('album-asin').value;
-//        const trackNumber = document.getElementById('track-number').value;
-//        const email = patient.id;
-//
-//        const songList = await this.client.addSongToPlaylist(email, asin, trackNumber, (error) => {
-//            errorMessageDisplay.innerText = `Error: ${error.message}`;
-//            errorMessageDisplay.classList.remove('hidden');
-//        });
-//
-//        this.dataStore.set('songs', songList);
-//
-//        document.getElementById('add-song').innerText = 'Add Song';
-//        document.getElementById("add-song-form").reset();
-//    }
 }
 
 /**
