--- web/static_assets/createPlaylist.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <script type="text/javascript" src="assets/createPlaylist.js"></script>
-    <link rel="stylesheet" type="text/css" href="css/style.css">
-</head>
-
-<body class="background">
-    <header class="header" id="header"></header>
-    <div class="card">
-        <h2>Create Playlist</h2>
-        <form class="card-content" id="create-playlist-form">
-            <p class="hidden error" id="error-message"> </p>
-            <p class="form-field">
-                <label>Playlist Name</label>
-                <input type="text" required class="validated-field" id="playlist-name" placeholder="smooth-jamz"
-                    autofocus>
-            </p>
-            <p class="form-field">
-                <label>Tags</label>
-                <input type="text" id="tags" class="tags" placeholder="tag1,tag2"
-                    data-tip="Must be a comma separated list of tags">
-            </p>
-            <p>
-                <a href="#" class="button" id="create">Create Playlist</a>
-            </p>
-        </form>
-    </div>
-</body>
-
-</html>
