+++ web/static_assets/index.html
@@ -2,26 +2,32 @@
 <html>
 
 <head>
-    <script type="text/javascript" src="assets/searchPlaylists.js"></script>
+    <script type="text/javascript" src="assets/home.js"></script>
     <link rel="stylesheet" type="text/css" href="css/style.css">
 </head>
 
 <body class="background">
+
     <header class="header" id="header"></header>
+
     <div class="card">
+
         <h2 class="heading">Employee Home</h2>
+
         <p class="container">
             <a href="newPrescription.html" class="button w50">New Prescription</a>
             <a href="searchPrescription.html" class="button w50">Search Prescription</a>
             <a href="newPatient.html" class="button w50">New Patient</a>
             <a href="searchPatient.html" class="button w50">Search Patient</a>
         </p>
+
     </div>
+
     <div class="card hidden" id='search-results-container'>
         <h3>Results for <span id="search-criteria-display"></span></h3>
-        <div id="search-results-display">
-        </div>
+        <div id="search-results-display"></div>
     </div>
+
 </body>
 
 </html>
