+++ web/static_assets/newPatient.html
@@ -7,41 +7,55 @@
 </head>
 
 <body class="background">
+
 <header class="header" id="header"></header>
+
 <div class="card">
+
     <h2 class="heading">New Patient</h2>
+
     <form class="card-content" id="create-patient-form">
+
         <p class="hidden error" id="error-message"> </p>
+
         <p class="form-field">
             <label>Patient Email</label>
-            <input type="email" required class="validated-field" id="email" placeholder="aoeuaoeu@aoeu.aoeu"
-                   autofocus>
+            <input type="email" required class="validated-field" id="email" placeholder="aoeuaoeu@aoeu.aoeu" autofocus>
         </p>
+
         <p class="form-field">
             <label>First Name</label>
             <input type="text" id="first-name" required class="tags" placeholder="Billy">
         </p>
+
         <p class="form-field">
             <label>Last Name</label>
             <input type="text" id="last-name" required class="tags" placeholder="Bob">
         </p>
+
         <p class="form-field">
             <label>Insurance</label>
             <input type="text" id="insurance" required class="tags" placeholder="Cigna">
         </p>
+
         <p class="form-field">
             <label>Phone</label>
             <input type="tel" id="phone" required class="tags" placeholder="123-456-7890">
         </p>
+
         <p class="form-field">
             <label>Address</label>
             <input type="text" id="address" required class="tags" placeholder="123 BillyBob Ln">
         </p>
+
         <p>
             <a href="#" class="button" id="create">Submit</a>
         </p>
+
     </form>
+
 </div>
+
 </body>
 
 </html>
