--- web/static_assets/playlist.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <script type="text/javascript" src="assets/viewPlaylist.js"></script>
-    <link rel="stylesheet" type="text/css" href="css/style.css">
-</head>
-
-<body class="background">
-    <header class="header" id="header"></header>
-    <div class="card">
-        <h2 id="playlist-name"></h2>
-        <h3>Created by: <span id="playlist-owner"></span></h3>
-        <div class="tags" id="tags"></div>
-        <ol id="songs"></ol>
-    </div>
-    <div class="card">
-        <h2>Add Song to Playlist</h2>
-        <form class="card-content" id="add-song-form">
-            <p class="hidden error" id="error-message"> </p>
-            <p class="form-field">
-                <label>Album Asin</label>
-                <input type="text" required class="validated-field" id="album-asin" placeholder="B00000000" autofocus>
-            </p>
-            <p class="form-field">
-                <label>Track Number</label>
-                <input type="number" id="track-number">
-            </p>
-        </form>
-        <a href="#" class="button" id="add-song">Add Song</a>
-    </div>
-</body>
-
-</html>
