+++ web/static_assets/prescription.html
@@ -2,11 +2,13 @@
 <html>
 
 <head>
+
     <script type="text/javascript" src="assets/viewPrescription.js"></script>
     <script type="text/javascript" src="assets/updatePrescription.js"></script>
     <script type="text/javascript" src="assets/deletePrescription.js"></script>
 
     <link rel="stylesheet" type="text/css" href="css/style.css">
+
 </head>
 
 <body class="background">
@@ -33,16 +35,6 @@
     <form class="card-content" id="update-prescription-form">
 
         <p class="hidden error" id="error-message"> </p>
-
-<!--        <p class="form-field">-->
-<!--            <label>Email</label>-->
-<!--            <input type="text" class="validated-field" id="email" placeholder="" autofocus readonly>-->
-<!--        </p>-->
-
-<!--        <p class="form-field">-->
-<!--            <label>Drug</label>-->
-<!--            <input type="text" class="validated-field" id="drug" placeholder="" autofocus readonly>-->
-<!--        </p>-->
 
         <p class="form-field">
             <label>Dose</label>
@@ -81,7 +73,6 @@
                 <option value="READY">Ready</option>
                 <option value="FILLED">Filled</option>
                 <option value="EXPIRED">Expired</option>
-<!--                <option value="DELETED">Deleted</option>-->
             </select>
         </p>
 
@@ -92,6 +83,5 @@
     <a href="#" class="button" id="delete-button">Delete</a>
 
 </div>
-
 
 </body>
