+++ web/static_assets/searchPatient.html
@@ -2,7 +2,6 @@
 <html>
 
 <head>
-<!--    <script type="text/javascript" src="assets/viewPatient.js"></script>-->
     <script type="text/javascript" src="assets/searchPatients.js"></script>
     <link rel="stylesheet" type="text/css" href="css/style.css">
 </head>
@@ -18,11 +17,9 @@
     <form class="card-content" id="search-patients-form">
 
         <p class="form-field">
-
             <label>Search by email</label>
             <input type="email" class="validated-field" id="email" placeholder="aoeuaoeu@aoeu.aoeu" autofocus>
             <sub>Search is case-sensitive and must match exactly.</sub>
-
         </p>
 
         <a href="#" class="button" id="search-single-btn">Go</a>
@@ -31,7 +28,6 @@
         <br>
 
         <p class="form-field">
-
             <br>
 
             <label>Search by last name or first and last name</label>
@@ -39,7 +35,6 @@
 
             <input type="text" class="validated-field" id="firstName" placeholder="First" autofocus>
             <sub>Search is case-sensitive and names must match exactly.</sub>
-
         </p>
 
             <a href="#" class="button" id="search-btn">Search</a>
