+++ web/static_assets/searchPrescription.html
@@ -17,11 +17,9 @@
     <form class="card-content" id="search-prescriptions-form">
 
         <p class="form-field">
-
             <label>Search by Prescription ID</label>
             <input type="text" class="validated-field" id="prescriptionId" placeholder="..." autofocus>
             <sub>Search is case-sensitive and must match exactly.</sub>
-
         </p>
 
         <a href="#" class="button" id="search-single-btn">Go</a>
@@ -30,7 +28,6 @@
         <br>
 
         <p class="form-field">
-
             <br>
 
             <label>Search by patient's email, status, or both</label>
@@ -46,14 +43,11 @@
             </select>
 
             <sub>Search is case-sensitive and must match exactly.</sub>
-
         </p>
 
         <a href="#" class="button" id="search-btn">Search</a>
 
     </form>
-
-<!--    <a href="#" class="button" id="create">Submit</a>-->
 
 </div>
